# docs/openapi.yaml
openapi: 3.1.0
info:
  title: UsageCheck Local API
  version: "1.0.0"
  description: |
    Read-only local HTTP API exposing the same Codex, Claude Code, and
    Antigravity (agy) usage that the UsageCheck tray menu displays.

    Designed so other coding agents (Codex, Claude Code, agy, Cursor, ...) can
    build MCP servers or skills on top of a stable JSON contract instead of
    scraping the tray UI.

    ## Semantics
    - All quota figures are **used percent** (`0` = unused, `100` = exhausted),
      matching the tray. Agy `remainingFraction` is already converted to used %.
    - Data is served from the latest background poll snapshot (refreshed every
      ~60s). The `updated_at` field reports when that snapshot was produced;
      it is `null` until the first poll completes.

    ## Security
    - The server binds to `127.0.0.1` only and is never exposed off-host.
    - No credentials, access tokens, or refresh tokens are ever returned.
    - The API is read-only (GET only).

    ## Enabling / configuring
    - Enabled by default when the tray app runs.
    - `USAGECHECK_API_DISABLE=1` disables the server.
    - `USAGECHECK_API_PORT=<port>` overrides the default port (`5178`).
servers:
  - url: http://127.0.0.1:5178
    description: Default local bind (override with USAGECHECK_API_PORT)
...
paths:
...
  /v1/usage/{provider}:
    get:
      tags: [usage]
      summary: One provider's usage
      description: Returns the usage snapshot filtered to a single provider.
      operationId: getUsageByProvider
      parameters:
        - name: provider
          in: path
          required: true
          description: Provider identifier.
          schema:
            $ref: "#/components/schemas/Provider"
      responses:
        "200":
          description: Usage snapshot for the requested provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageResponse"
        "404":
          description: Unknown provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: unknown_provider
                message: "unknown provider 'foo' (expected codex, claude, or agy)"
...
components:
  schemas:
    Provider:
      type: string
      enum: [codex, claude, agy, cursor, grok, higgsfield]
      description: Usage provider category.
    Status:
      type: string
      enum: [ok, needs_login, rate_limited, error, stale, waiting_for_usage, identity_changed, experimental_error, needs_setup]
      description: |
        Freshness/auth status for an account's most recent poll.
        - `ok`: live data fetched successfully.
        - `needs_login`: token missing/expired (401/403); may fall back to local logs.
        - `rate_limited`: provider returned 429.
        - `error`: network/other failure.
        - `stale`: transient failure; serving the last successful in-memory result.
        - `waiting_for_usage`: authenticated but no usage windows reported yet.
        - `identity_changed`: the CLI profile now belongs to a different identity than expected.
        - `experimental_error`: an experimental (reference/DB) provider probe failed.
        - `needs_setup`: provider connected but not yet configured (e.g. Higgsfield credits absent).
    Quota:
      type: object
      description: A single usage window (used percent).
      required: [used_percent]
      properties:
        used_percent:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: Percent of the window consumed (0-100).
          example: 38.0
        window_label:
          type: string
          description: Human label derived from window_seconds (e.g. "5h", "7d").
          example: "5h"
        window_seconds:
          type: [integer, "null"]
          format: int64
          description: Length of the quota window in seconds, when known.
          example: 18000
        resets_at:
          type: [string, "null"]
          format: date-time
          description: RFC3339 timestamp when the window resets, when known.
          example: "2026-07-10T08:41:00Z"
    Pool:
      type: object
      description: |
        An Antigravity (agy) Model Quota pool. Agy exposes two shared pools:
        "Gemini Models" and "Claude and GPT models".
      required: [name]
      properties:
        name:
          type: string
          example: "Gemini Models"
        five_hour:
          oneOf:
            - $ref: "#/components/schemas/Quota"
            - type: "null"
        week:
          oneOf:
            - $ref: "#/components/schemas/Quota"
            - type: "null"
    TokenTotals:
      type: object
      description: |
        Local-log token totals (Codex/Claude fallback only, when the live quota
        API is unavailable). Zero for accounts served from the live quota API.
      required: [five_hours, week, month]
      properties:
        five_hours:
          type: integer
          format: int64
          example: 0
        week:
          type: integer
          format: int64
          example: 0
        month:
          type: integer
          format: int64
          example: 0
    AccountUsage:
      type: object
      required: [id, provider, display_name, status, pools, token_totals]
      properties:
        id:
          type: string
          description: Stable account identifier (UUID).
          example: "6f1c2b8e-0a1d-4c33-9f6e-2b7c9d0e1a22"
        provider:
          $ref: "#/components/schemas/Provider"
        display_name:
          type: string
          description: Email or plan-aware account name.
          example: "user@example.com"
        plan:
          type: [string, "null"]
          description: Provider plan name, when known.
          example: "pro"
        status:
          $ref: "#/components/schemas/Status"
        five_hour:
          oneOf:
            - $ref: "#/components/schemas/Quota"
            - type: "null"
          description: Account-level 5-hour window (most constrained pool for agy).
        week:
          oneOf:
            - $ref: "#/components/schemas/Quota"
            - type: "null"
          description: Account-level 7-day window (most constrained pool for agy).
        pools:
          type: array
          description: Per-pool breakdown (agy only; empty for Codex/Claude).
          items:
            $ref: "#/components/schemas/Pool"
        token_totals:
          $ref: "#/components/schemas/TokenTotals"
        local_status:
          type: [string, "null"]
          description: >-
            Provenance of the locally-aggregated token usage when it is not a
            clean `ok` (e.g. `unavailable`, `partial`, `truncated`, `ambiguous`,
            `conflict`, `assumed`, `no_local_profile`). Absent/null means the
            local totals are fully trustworthy. Lets clients distinguish a real
            zero from a failed/ambiguous local scan.
          example: "assumed"
        detail_suffix:
          type: [string, "null"]
          description: Provider-specific trailing detail, e.g. remaining API credits.
    UsageResponse:
      type: object
      required: [updated_at, count, accounts]
      properties:
        updated_at:
          type: [string, "null"]
          format: date-time
          description: When the served snapshot was produced (null before first poll).
          example: "2026-07-10T03:41:00Z"
        count:
          type: integer
          description: Number of accounts in `accounts`.
          example: 3
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountUsage"
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: unknown_provider
        message:
          type: string
          example: "unknown provider 'foo' (expected codex, claude, or agy)"